requirements-docker-cpu.txt
# Requirements for Docker CPU build
# (torch/torchvision installed separately with CPU-only)

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
pillow>=10.0.0
mlflow>=2.8.0